Add custom node allocator
Using free/malloc for the accumulator nodes (especially for pollard)
turned out to be slower than a custom allocator that can only allocate
objects of a fixed size. Not sure if this is worth the extra code
complexity. It would fit in nicely with the optimal caching strategy
because it pre allocates the exact number of nodes needed.