accumulator: Introduce Read() virtual method
The Read() method is defined in the accumulator class as a pure virtual
function. This makes it easier to define shared functions such as a
shared Prove() method.

Note that the ram_forest will always return a value so it's ok to directly
call value() for any reads for ram_forest.